Fully implement MSAL authentication library
